+++ CHANGELOG.md
@@ -133,6 +133,7 @@ Estos son los cambios hechos en la ECORED-BASE-2021
 - /src/views/Tema1.vue
 - /src/views/Tema2.vue
 - /src/views/Tema3.vue
+- /src/views/Tema4.vue
 
 ### Agregado